+++ client/src/App.jsx
@@ -38,13 +38,13 @@ import {
   FooterNavbar,
   RenkuNavBar,
 } from "./features/landing/components/NavBar/NavBar";
-import {
-  LegacyDatasetAddToProject,
-  LegacyDatasets,
-  LegacyProjectView,
-  LegacyRoot,
-  LegacyShowDataset,
-} from "./features/legacy";
+
+import LegacyDatasets from "./features/legacy/datasets/LegacyDatasets";
+import LegacyDatasetAddToProject from "./features/legacy/datasets/LegacyDatasetAddToProject";
+import LegacyRoot from "./features/legacy/LegacyRoot";
+import LegacyProjectView from "./features/legacy/LegacyProjectView";
+import LegacyShowDataset from "./features/legacy/LegacyShowDataset";
+
 import LoginHandler from "./features/loginHandler/LoginHandler";
 import { Unavailable } from "./features/maintenance/Maintenance";
 import LazyRootV2 from "./features/rootV2/LazyRootV2";
