+++ client/src/features/legacy/LegacyShowDataset.tsx
@@ -1,0 +1,52 @@
+/*!
+ * Copyright 2025 - Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { useContext } from "react";
+
+import { DatasetCoordinator } from "../../dataset/Dataset.state";
+import LazyShowDataset from "../../dataset/LazyShowDataset";
+import AppContext from "../../utils/context/appContext";
+
+import type { UserInfo } from "../usersV2/api/users.types";
+
+import NoLegacySupport from "./NoLegacySupport";
+
+interface LegacyDatasetProps {
+  userInfo: UserInfo;
+}
+
+export default function LegacyShowDataset({ userInfo }: LegacyDatasetProps) {
+  const { client, model: contextModel, params } = useContext(AppContext);
+  if (params && !params.LEGACY_SUPPORT.enabled) {
+    return <NoLegacySupport />;
+  }
+
+  const model = contextModel as { subModel: (arg0: string) => unknown };
+  return (
+    <LazyShowDataset
+      insideProject={false}
+      client={client}
+      projectsUrl="/projects"
+      datasetCoordinator={
+        new DatasetCoordinator(client, model.subModel("dataset"))
+      }
+      logged={userInfo?.isLoggedIn ?? false}
+      model={model}
+    />
+  );
+}
