--- client/src/features/legacy/index.tsx
@@ -1,7 +1,0 @@
-export {
-  LegacyDatasetAddToProject,
-  LegacyDatasets,
-  LegacyProjectView,
-  LegacyRoot,
-  LegacyShowDataset,
-} from "./legacy";
